Week 1
Overall, week 1 was just an overview of what the course will look like over the next couple weeks. We were introduced to some of the authors of the textbook which was really cool to see because we got some in site on why they made the course a certain way. I really liked Guido van Rossum who is the creator of Python.